# swagger/swagger.yaml
swagger: '2.1'
info:
  title: Docker Vulnerability Manager API
  description: Docker containers' source code freshness management API
  version: 0.19.0
consumes:
  - application/json
produces:
  - application/json
basePath: /api
components:
  parameters:
    token:
      in: header
      name: token
      required: true
      schema:
        type: string
      description: authentication token
      example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjVhOTVhNzBiYjljMjQ1MzA2YjE2YzE0NSIsImlhdCI6MTUxOTkyMzc5NiwiZXhwIjoxNTE5OTI3Mzk2fQ.e5cKEEGtDb54AVlnZHf-P132x1LTrS4WrgXywfmkSkk
...
    imageIdPath:
      in: path
      name: imageId
      required: true
      schema:
        type: string
      description: ID of a Docker image
...
    imageName:
      in: body
      name: imageName
      required: true
      schema:
        type: string
      description: name of the Docker image
      example: pawelpaszki/vuln-demo-2-node
...
paths:
  /containers:
    get:
      summary: List all containers.
      description:  Lists all of the Docker containers on the host
      operationId: listContainers
      parameters:
        - $ref: '#/components/parameters/token'
      tags:
        - Containers
      responses:
        '200':
          description: OK
          schema:
            type: object
            properties:
              Id:
                type: string
                example: fe9593f590c180f972d1a9f2a29c49b76c0c61c83919fe362b53abb976f5a94a
              Names:
                type: array
                example: ["/nifty_kowalevski"]
              Image:
                type: string
                example: pawelpaszki/vuln-demo-2-node
              ImageID:
                type: string
                example: sha256:d8217af59a1d6dfd795d50b67e9df4a4785bdd4d4b56a98942a5c06a1a67a0fa
              Command:
                type: string
                example: npm start
              Created:
                type: date
                example: 1519669038
              Ports:
                type: array
                example: []
              Labels:
                type: object
                example: {}
              State:
                type: string
                example: Created
              Status:
                type: string
                example: Created
              HostConfig:
                type: object
                example: { "NetworkMode": "default" }
              NetworkSettings:
                type: array
                example: []
              Mounts:
                type: array
                example: []
  /containers/create:
    post:
      summary: Create a container
      description: create new Docker container
      operationId: createContainer
      tags:
        - Containers
      parameters:
        - $ref: '#/components/parameters/token'
        - $ref: '#/components/parameters/imageName'
      responses:
        '201':
          description: OK
          schema:
            type: object
            properties:
              id:
                type: string
                example: fe9593f590c180f972d1a9f2a29c49b76c0c61c83919fe362b53abb976f5a94a
        '500':
          description: Unable to create container
...
  /images/search:
    post:
      summary: Search for Docker images
      description: Search for Docker images available in Docker registry
      operationId: searchImages
      tags:
        - Images
      parameters:
        - $ref: '#/components/parameters/token'
        - $ref: '#/components/parameters/imageName'
      responses:
        '200':
          description: OK
          schema:
            type: array
            example: [
                       "ubuntu",
                       "dorowu/ubuntu-desktop-lxde-vnc",
                       "rastasheep/ubuntu-sshd",
                       "ansible/ubuntu14.04-ansible",
                       "ubuntu-upstart",
                       "neurodebian",
                       "ubuntu-debootstrap",
                       "1and1internet/ubuntu-16-nginx-php-phpmyadmin-mysql-5",
                       "nuagebec/ubuntu",
                       "tutum/ubuntu",
                       "ppc64le/ubuntu",
                       "i386/ubuntu",
                       "1and1internet/ubuntu-16-apache-php-7.0",
                       "eclipse/ubuntu_jdk8",
                       "1and1internet/ubuntu-16-apache",
                       "codenvy/ubuntu_jdk8",
                       "darksheer/ubuntu",
                       "1and1internet/ubuntu-16-nginx-php-5.6-wordpress-4",
                       "1and1internet/ubuntu-16-nginx-php-phpmyadmin-mariadb-10",
                       "pivotaldata/ubuntu",
                       "1and1internet/ubuntu-16-healthcheck",
                       "1and1internet/ubuntu-16-sshd",
                       "ossobv/ubuntu",
                       "pivotaldata/ubuntu-gpdb-dev",
                       "smartentry/ubuntu"
                   ]
...
  /images/checktag:
    post:
      summary: get tag
      description: get latest Image's tag
      operationId: checkTag
      tags:
        - Images
      parameters:
        - $ref: '#/components/parameters/token'
        - $ref: '#/components/parameters/imageName'
      responses:
        '200':
          description: OK
          schema:
            type: object
            properties:
              major:
                type: string
                example: 0
              minor:
                type: string
                example: 0
              patch:
                type: string
                example: 1
...
  /images/{id}:
    delete:
      summary: Delete docker image
      description: Delete existing Docker image from the host
      operationId: deleteImage
      tags:
        - Images
      parameters:
        - $ref: '#/components/parameters/token'
        - $ref: '#/components/parameters/imageIdPath'
      responses:
        '200':
          description: OK
          schema:
            type: object
            properties:
              message:
                type: string
                example: Image removed successfully
        '404':
          description: Unable to remove container
          schema:
            type: object
            properties:
              error:
                type: string
                example: Image not found
        '409':
          description: Image cannot be removed
          schema:
            type: object
            properties:
              error:
                type: string
                example: Image cannot be removed
...